RavenDB-18879 - fixing PR comments#4
