+++ src/Raven.Server/Documents/Sharding/Handlers/Processors/Replication/ShardedReplicationHandlerProcessorForGetTombstones.cs
@@ -51,6 +51,11 @@ namespace Raven.Server.Documents.Sharding.Handlers.Processors.Replication
             {
                 var final = new GetTombstonesPreviewResult();
 
+                foreach (var result in results.Span)
+                {
+                    result.Tombstones.Sort(TombstonesPreviewComparer.Instance);
+                }
+
                 final.Tombstones = _handler.DatabaseContext.Streaming.PagedShardedItem(
                     results,
                     selector: r => r.Tombstones,
